Refactored based on review
