+++ actions/list_sites.py
@@ -16,7 +16,7 @@ def list_sites(api_auth, parameters):
     org = parameters["organisation"]
 
     # Get all sites and return a response based on the number of sites
-    res = api_auth.api_site.list_sites()
+    res = api_auth.site.list_sites()
 
     if res.status_code == 200:
         data = res.json()["items"]
