+++ actions/list_sites_followup.py
@@ -28,7 +28,7 @@ def list_sites_followup(api_auth, parameters):
 
 
     # Get all sites and return a response based on the number of sites
-    res = api_auth.api_site.list_sites()
+    res = api_auth.site.list_sites()
 
     if res.status_code == 200:
         data = res.json()["items"]
