+++ tests/test_clear_sites.py
@@ -6,7 +6,7 @@ import app
 
 from flask import json
 from samples.create_site import *
-from api.api_auth import SteelConnectAPI
+from api import SteelConnectAPI
 
 
 # class MockResponse(Mock):
